% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EUdata}
\alias{EUdata}
\title{Inflation and Unemployment Dataset}
\format{
An array of dimension 3 x 11 x 250
}
\source{
Canova and Ciccarelli (2009) DOI: 10.1111/j.1468-2354.2009.00554.x
}
\usage{
EUdata
}
\description{
This dataset contains monthly macroeconomic variables for 11 EU countries
(Austria, Belgium, Estonia, Finland, France, Germany, Greece, Ireland, Italy,
The Netherlands, Portugal, and Spain) from January 2002 to October 2022. The
variables include the Industrial Production Index, Price Index, and
Unemployment Rate, sourced from FRED and EUROSTAT. Series are adjusted for
seasonality and working days, and stationarity is enforced via differencing.
}
\details{
Outliers are detected in key periods such as the COVID-19 pandemic (Mar–Aug 2020, Nov 2020),
the Ukrainian conflict (Mar–Apr 2022), and the inflation rise (Sep–Oct 2022).
}
\keyword{datasets}
