% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tradedata}
\alias{Tradedata}
\title{Trade Network Dataset}
\format{
An array of dimension 27 x 27 x 22
}
\source{
Rose (2004) DOI: 10.1257/000282804322970724
}
\usage{
Tradedata
}
\description{
This dataset comprises 22 annual trade networks among 27 countries from 1995
to 2017, derived from IMF reports and other official sources (COMTRADE,
EUROSTAT). Trade values are transformed to constant USD (base year 2010) using
the US GDP PPP deflator. The selected countries are large economies, filtered
for data completeness.
}
\details{
Notable outliers are detected in 2015 and 2017, while 2016 is not flagged as
anomalous.
}
\keyword{datasets}
