% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Volatilitydata}
\alias{Volatilitydata}
\title{Volatility Network Dataset}
\format{
An array of dimension 50 x 50 x 145
}
\source{
Billio, Casarin, and Iacopini (2022) DOI: 10.1080/07350015.2022.2032721
}
\usage{
Volatilitydata
}
\description{
This dataset consists of 145 weekly volatility networks (Monday–Friday) from
January 2016 to September 2020, constructed using pairwise Granger causality
for 50 firms from Germany, France, and Italy. Firms span 11 GICS sectors,
including Financials, IT, Industrials, Health Care, and others.
}
\details{
Outliers capture volatility regime changes before and during the COVID-19
crisis, particularly transitions such as from "low" to "high" and "moderate"
volatility levels.
}
\keyword{datasets}
